+++ modules/client/examples/06-multisig-client/12-get-members.ts
@@ -5,7 +5,7 @@ import {
   Context,
   ContextPlugin,
   MultisigClient,
-  MultisigPluginSettings,
+  MultisigVotingSettings,
 } from "@aragon/sdk-client";
 import { contextParams } from "../00-client/00-context";
 
@@ -18,21 +18,13 @@ const client = new MultisigClient(contextPlugin);
 
 const daoAddressorEns = "0x12345...";
 
-const settings: MultisigPluginSettings = await client.methods
-  .getPluginSettings(daoAddressorEns);
+const settings: string[] = await client.methods
+  .getMembers(daoAddressorEns);
 console.log(settings);
 /*
-{
-  members: [
-    "0x1234567890123456789012345678901234567890",
-    "0x2345678901234567890123456789012345678901",
-    "0x3456789012345678901234567890123456789012",
-    "0x4567890123456789012345678901234567890123",
-    "0x5678901234567890123456789012345678901234",
-  ],
-  votingSettings: {
-    minApprovals: 4,
-    onlyListed: true
-  }
-}
+[
+  "0x1234567890...",
+  "0x2345678901...",
+  "0x3456789012...",
+]
 */