+++ modules/client/src/multisig/internal/graphql-queries/members.ts
@@ -1,0 +1,11 @@
+import { gql } from "graphql-request";
+
+export const QueryMultisigMembers = gql`
+query MultisigMembers($address: ID!) {
+    multisigPlugin(id: $address){
+        members {
+            address
+        }
+    }
+}
+`;